backend/requirements.txt
flask>=2.2.3
flask-cors>=3.0.10
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.2.2
joblib>=1.2.0
gunicorn>=20.1.0
python-dotenv>=1.0.0
supabase>=1.0.3
psycopg2-binary>=2.9.6
azure-storage-blob>=12.14.0
azure-identity>=1.12.0 